CORE-8743 Refactor sharing presenter. Call sharing service.
