+++ de-lib/src/main/java/org/iplantc/de/apps/client/views/tools/ToolSharingDialog.java
@@ -4,20 +4,26 @@
 
 package org.iplantc.de.apps.client.views.tools;
 
+import org.iplantc.de.apps.client.gin.factory.ToolSharingPresenterFactory;
+import org.iplantc.de.client.models.tool.Tool;
 import org.iplantc.de.commons.client.views.dialogs.IPlantDialog;
 
 import com.google.inject.Inject;
 
 import com.sencha.gxt.widget.core.client.event.SelectEvent;
 
+import java.util.List;
+
 public class ToolSharingDialog extends IPlantDialog {
 
-    @Inject
+
     ToolSharingPresenter presenter;
+    ToolSharingPresenterFactory factory;
     
     @Inject
-    ToolSharingDialog() {
+    ToolSharingDialog(ToolSharingPresenterFactory factory) {
         super(false);
+        this.factory = factory;
         setPixelSize(600, 500);
         setHideOnButtonClick(true);
         setModal(true);
@@ -30,6 +36,17 @@ public class ToolSharingDialog extends IPlantDialog {
               presenter.processRequest();
             }
         });
+    }
+
+    public void show(final List<Tool> resourcesToShare) {
+        presenter = factory.create(resourcesToShare);
+        presenter.go(this);
+        super.show();
+    }
+
+    @Override
+    public void show() throws UnsupportedOperationException {
+        throw new UnsupportedOperationException("This method is not supported for this class. ");
     }
 
 }
