+++ de-lib/src/main/java/org/iplantc/de/client/models/tool/sharing/ToolPermission.java
@@ -1,15 +1,19 @@
 package org.iplantc.de.client.models.tool.sharing;
 
-import org.iplantc.de.client.models.HasId;
+import com.google.web.bindery.autobean.shared.AutoBean;
 
 /**
  * Created by sriram.
  */
-public interface ToolPermission extends HasId {
+public interface ToolPermission {
 
     void setPermission(String permission);
 
     String getPermission();
 
+    @AutoBean.PropertyName("tool_id")
     void setId(String id);
+
+    @AutoBean.PropertyName("tool_id")
+    String getId();
 }
