+++ de-lib/src/main/java/org/iplantc/de/client/services/ToolServices.java
@@ -1,12 +1,9 @@
 package org.iplantc.de.client.services;
 
-import org.iplantc.de.apps.client.presenter.tools.ManageToolsViewPresenter;
 import org.iplantc.de.client.models.tool.Tool;
 import org.iplantc.de.client.models.tool.sharing.ToolSharingRequestList;
 import org.iplantc.de.client.models.tool.sharing.ToolUnSharingRequestList;
 import org.iplantc.de.shared.AppsCallback;
-
-import com.google.gwt.user.client.rpc.AsyncCallback;
 
 import com.sencha.gxt.data.shared.loader.FilterPagingLoadConfig;
 
@@ -14,17 +11,17 @@ import java.util.List;
 
 public interface ToolServices {
 
-    void searchTools(FilterPagingLoadConfig loadConfig, AsyncCallback<List<Tool>> callback);
+    void searchTools(FilterPagingLoadConfig loadConfig, AppsCallback<List<Tool>> callback);
 
-    void getTools(AsyncCallback<List<Tool>> callback);
+    void getTools(AppsCallback<List<Tool>> callback);
 
-    void addTool(Tool tool, AsyncCallback<Tool> callback);
+    void addTool(Tool tool, AppsCallback<Tool> callback);
 
-    void deleteTool(Tool tool, AsyncCallback<String> callback);
+    void deleteTool(Tool tool, AppsCallback<String> callback);
 
-    void getPermissions(List<Tool> currentSelection, AsyncCallback<String> callback);
+    void getPermissions(List<Tool> currentSelection, AppsCallback<String> callback);
 
-    void shareTool(ToolSharingRequestList obj, AsyncCallback<String> appsCallback);
+    void shareTool(ToolSharingRequestList obj, AppsCallback<String> callback);
 
-    void unShareTool(ToolUnSharingRequestList obj, AsyncCallback<String> appsCallback);
+    void unShareTool(ToolUnSharingRequestList obj, AppsCallback<String> callback);
 }
