+++ de-lib/src/main/java/org/iplantc/de/client/services/converters/ToolsCallbackConverter.java
@@ -3,18 +3,18 @@ package org.iplantc.de.client.services.converters;
 import org.iplantc.de.client.models.tool.Tool;
 import org.iplantc.de.client.models.tool.ToolAutoBeanFactory;
 import org.iplantc.de.client.models.tool.ToolList;
+import org.iplantc.de.shared.AppsCallback;
 
-import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.web.bindery.autobean.shared.AutoBean;
 import com.google.web.bindery.autobean.shared.AutoBeanCodex;
 
 import java.util.List;
 
-public class ToolsCallbackConverter extends AsyncCallbackConverter<String, List<Tool>> {
+public class ToolsCallbackConverter extends DECallbackConverter<String, List<Tool>> {
 
     private final ToolAutoBeanFactory factory;
 
-    public ToolsCallbackConverter(AsyncCallback<List<Tool>> callback, ToolAutoBeanFactory factory) {
+    public ToolsCallbackConverter(AppsCallback<List<Tool>> callback, ToolAutoBeanFactory factory) {
         super(callback);
         this.factory = factory;
     }
